requirements.txt
python-dotenv
FastAPI
uvicorn[standard]